add endpoint list user goals, a user summary, goal detail by user

+++ src/BackEnd/BpChallenge/BpChallenge.Domain/Entities/FundingShareValue.cs
@@ -4,7 +4,7 @@ namespace BpChallenge.Domain.Entities;
 
 public class FundingShareValue : BaseEntity
 {
-    public decimal Value { get; private set; }
+    public double Value { get; private set; }
     public DateTime Date { get; private set; }
     public int FundingId { get; private set; }
     public Funding Funding { get; private set; }
